Don't make unnecessary (and condescending) noise in offline mode
"When a program has nothing surprising to say, it should say nothing."

-- Rule of Silence, The Art of Unix Programming, Eric Steven Raymond
   http://www.catb.org/~esr/writings/taoup/html/ch01s06.html#id2878450

Given that offline mode is enabled only by a command-line argument,
spitting out a message telling the user that offline mode is enabled
is redundant.

Making the message a warning is senseless.  The user doesn't need to
be warned that their instructions have been carried out, only when
their instructions fail.

Asking users to check for updates regularly is condescending and
offensive.