as per https://github.com/LedFx/LedFx/issues/235
rain effect has concurrent access risk of the render data from two seperate threads

Calls into numpy np.flatnonzero(self.drop_frames) does an integrity check on the data and will panic if the data is changed between enter and exit

Such concurrency can cause other unexpected side effects, even if not trapped by numpy

Therefore the two paths should run an threading lock, this is substantiated in the base class of effect(...)

Then calls into the child implementations of render and audio_data_update use the lock to protect code execution that may update or use the render data structs, understanding that this may not be specific to the rain effect

It is worth noting that a pre existing naming convention of _audio_data_updated calls into audio_data_updated, I would expect the _ to be the internal private function, but in this case it is the publically called. I have followed this naming convention for wrappign render with _render, where it did not otherwise have wrapping code

This fix in various versions has run for hours without failure, prior MTBF was ~15 min for rain, sometime much quicker, as it is purely chance.