Seeing crashes on startup with multiple WLED controllers, now running 4 x athom and 1 x pixel stick
Stack trace below, which is a trap to check that the pixel render frame is the same size as the subdevice

        if data.size != self._config["channel_count"]:

However, could reproduce what looked like one device using the config of another.

This was actually due to code in setup_subdevice which is a reference, not a copy to the class level single instance variable DEVICE_CONFIGS

config = self.DEVICE_CONFIGS[self._config["sync_mode"]]

as later code makes changes to the content of config, its a shared structure across all instances of the class, and therefore each registered device can damage others. I don't know how this worked at all once I was running devices with different frame sizes.

The simplest fix is to move DEVICE_CONFIGS into the __init__ method, so it is a seperate variable every time.
Otherwise a deepcopy could be used. Will leave that to review comments.

Other two structures at class level are effectively constants. Maybe it was the original coders intent that DEVICE_CONFIGS be a constant as well, but handling by reference causes the problem. Not sure if we should change the naming to device_configs as it is not a constant now.

Exception in thread Thread-6 (thread_function):
Traceback (most recent call last):
  File "C:\Users\atod\AppData\Local\Programs\Python\Python310\lib\threading.py", line 1009, in _bootstrap_inner
[INFO    ] ledfx.config                   : Saving configuration file to C:\Users\atod\AppData\Roaming.ledfx
    self.run()
  File "C:\Users\atod\AppData\Local\Programs\Python\Python310\lib\threading.py", line 946, in run
    self._target(*self._args, **self._kwargs)
  File "C:\Users\atod\PycharmProjects\LedFxbrf\ledfx\virtuals.py", line 374, in thread_function
    self.flush()
  File "C:\Users\atod\PycharmProjects\LedFxbrf\ledfx\virtuals.py", line 531, in flush
    device.update_pixels(self.id, data)
  File "C:\Users\atod\PycharmProjects\LedFxbrf\ledfx\devices__init__.py", line 143, in update_pixels
    self.flush(frame)
  File "C:\Users\atod\PycharmProjects\LedFxbrf\ledfx\devices\wled.py", line 106, in flush
    self.subdevice.flush(data)
  File "C:\Users\atod\PycharmProjects\LedFxbrf\ledfx\devices\e131.py", line 123, in flush
    raise Exception(
Exception: Invalid buffer size. 1536 != 1518