refactor: rename device_repeat to pixels_per_device for clarity (#1323)
* refactor: rename device_repeat to pixels_per_device for clarity

Imho pixels_per_device is a much better name since it describes what this setting is doing. With device_repeat I expect that the pixels are repeated for every device and device_repeat is specifying the number of devices.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* update old configs

* Fix migration

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Delete old device_repeat entry from config after migration

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* ci: ubuntu: Fix build with cmake 4.0

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>

Merging with mitigation for 

https://github.com/actions/runner-images/issues/11926

Will revisit when this shakes out