Feat: add a `active` flag to scenes based on current state matching scene definition
In some cases (automation frequently) - it's nice to know that a given scene is active.

This accomplishes that by adding an `active` flag to the `GET /api/scenes` endpoint that returns true or false based on whether or not the scene config matches the current state of the virtual devices.

Added a basic test