Feat: add an `active` flag to scenes based on current state of devices matching scene definition (#1568)
* Feat: add a `active` flag to scenes based on current state matching scene definition

In some cases (automation frequently) - it's nice to know that a given scene is active.

This accomplishes that by adding an `active` flag to the `GET /api/scenes` endpoint that returns true or false based on whether or not the scene config matches the current state of the virtual devices.

Added a basic test

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* feat: add docs for `active`

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Spelling for api docs

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Scene active true clarity

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>